use size_t for GSL MC funciton signature
git-svn-id: https://root.cern.ch/svn/root/branches/dev/mathDev@20149 27541ba8-7e3a-0410-8455-c3a389f83636

+++ mathmore/inc/Math/GSLMCIntegrator.h
@@ -156,7 +156,7 @@ namespace Math {
          
       void SetFunction(const IMultiGenFunction &f); 
    
-      typedef double ( * GSLMonteFuncPointer ) ( double *, unsigned int, void *);    
+      typedef double ( * GSLMonteFuncPointer ) ( double *, size_t, void *);    
       
       void SetFunction( GSLMonteFuncPointer f, void * p = 0);
